cleaning stuff out before actually working
added the crusader, which is not the final name, nor the latest version of it. ive been working on the stand part in a separate file.
sage gets a few infrastructure things, still nonfunctional
fixed andromeda launcher creator
minor change on the arc classique, still needs to get its functionality